# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Partial 1 - Rugby World Cup 2023'","description":"'Partial 1 - Rugby World Cup 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Partial 1 - Rugby World Cup 2023'"}],"paths":{"/rwc/2023/pools":{"get":{"tags":["rugby-controller"],"operationId":"GetAllPools","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PoolResponseDTO"}}}}}}}},"/rwc/2023/pools/{pool_id}":{"get":{"tags":["rugby-controller"],"operationId":"GetPoolById","parameters":[{"name":"pool_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PoolResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"PoolResponseDTO":{"type":"object","properties":{"pool_id":{"type":"string"},"teams":{"type":"array","items":{"$ref":"#/components/schemas/TeamDTO"}}}},"TeamDTO":{"type":"object","properties":{"team_id":{"type":"integer","format":"int64"},"team_name":{"type":"string"},"country":{"type":"string"},"matches_played":{"type":"integer","format":"int32"},"wins":{"type":"integer","format":"int32"},"draws":{"type":"integer","format":"int32"},"losses":{"type":"integer","format":"int32"},"points_for":{"type":"integer","format":"int32"},"points_against":{"type":"integer","format":"int32"},"points_differential":{"type":"integer","format":"int32"},"tries_made":{"type":"integer","format":"int32"},"bonus_points":{"type":"integer","format":"int32"},"points":{"type":"integer","format":"int32"},"total_yellow_cards":{"type":"integer","format":"int32"},"total_red_cards":{"type":"integer","format":"int32"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}